:truck: refactor: move entity2 to entity, removed old code
