--- src/entity/adapters/generateId.test.ts
@@ -1,11 +1,0 @@
-import { describe, it, expect, expectTypeOf } from "vitest"
-import { generateId } from "./generateId"
-
-describe("generateId", () => {
-  it("Given generateId function, When called, Then return string", () => {
-    const result = generateId()
-
-    expect(result).not.toBeUndefined()
-    expectTypeOf(result).toEqualTypeOf<string>()
-  })
-})
