--- src/entity/adapters/generateId.ts
@@ -1,5 +1,0 @@
-import { createId } from "@paralleldrive/cuid2"
-
-export function generateId(): string {
-  return createId()
-}
