+++ src/entity/interface/data.test.ts
