+++ src/entity/interface/data.ts
