+++ src/entity/interface/entity.test.ts
