+++ src/entity/interface/entity.ts
