+++ src/entity/interface/index.ts
