+++ src/entity/interface/proxyTarget.ts
