+++ src/entity/interface/sync.ts
