+++ src/entity/proto.test.ts
