+++ src/entity/proto.ts
