+++ src/entity/sync.ts
@@ -1,40 +1,20 @@
-import { Opaque } from "type-fest"
+import { SyncTarget, SyncKey } from "./interface"
 
-export type SyncKey = Opaque<string, "sync-key">
-
-export enum SyncStatus {
-  NOT_SYNCED = "not-synced",
-  SYNCING = "syncing",
-  SYNCED = "synced",
-  ERROR = "error",
-}
-
-export type SyncDestination = {
-  id: string | symbol
-  status: SyncStatus
-  lastSyncedAt?: Date
-}
-
-function makeSyncDestination(id: string | symbol): SyncDestination {
-  return {
-    id,
-    status: SyncStatus.NOT_SYNCED,
-    lastSyncedAt: undefined,
-  }
+export function makeSyncKey(id: string): SyncKey {
+  return id as SyncKey
 }
 
 export class SyncMap {
-  private _map: Map<SyncKey, SyncDestination>
-
-  public static makeSyncKey(id: string): SyncKey {
-    return id as SyncKey
-  }
+  private _map: Map<SyncKey, SyncTarget>
 
   constructor(syncIds: SyncKey[]) {
     this._map = new Map()
 
     syncIds.forEach((id) => {
-      this._map.set(id, makeSyncDestination(id))
+      this._map.set(id, {
+        id,
+        status: "not-synced",
+      })
     })
   }
 
@@ -49,7 +29,7 @@ export class SyncMap {
       return false
     }
 
-    return destination.status === SyncStatus.SYNCED
+    return destination.status === "synced"
   }
 
   public setStatus(id: SyncKey, promise: Promise<unknown>): void {
@@ -58,13 +38,13 @@ export class SyncMap {
     if (destination) {
       promise
         .then((resolvedValue: unknown) => {
-          destination.status = SyncStatus.SYNCED
+          destination.status = "synced"
           destination.lastSyncedAt = new Date()
 
           return resolvedValue
         })
         .catch(() => {
-          destination.status = SyncStatus.ERROR
+          destination.status = "error"
         })
     }
   }
