+++ src/entity/types.ts
@@ -1,4 +1,4 @@
-export type DeepReadonly<TData extends Record<string, any>> = {
+export type DeepReadonly<TData> = {
   readonly [Key in keyof TData]: TData[Key] extends Record<string, any>
     ? TData[Key] extends Set<any> | Map<any, any> | Date | RegExp
       ? TData[Key]
